refactor: move months.tsx to months/months.tsx
- Move src/components/certificates/months.tsx to
  src/components/certificates/months/months.tsx
- Update the import paths to this file in certificates.tsx and
  certificates-preview.tsx
