+++ src/components/certificates/certificates.tsx
@@ -1,5 +1,5 @@
 import { ChangeEvent, useState } from "react";
-import { Months } from "./months";
+import { Months } from "./months/months";
 import { Years } from "./years";
 import { CheckBox } from "./check-box";
 import { useForm } from "react-hook-form";
